Cleans up authorization + adds some tests
