+++ godmin.gemspec
@@ -22,7 +22,6 @@ Gem::Specification.new do |s|
   s.add_dependency "bcrypt", "~> 3.1.7"
   s.add_dependency "bootstrap-sass", "~> 3.1.1.0"
   s.add_dependency "kaminari", "~> 0.15.1"
-  s.add_dependency "pundit", "~> 0.3.0"
   s.add_dependency "select2-rails", "~> 3.5.7"
   s.add_dependency "simple_form", "~> 3.0.0"
