+++ lib/godmin.rb
@@ -1,6 +1,5 @@
 require "bootstrap-sass"
 require "kaminari"
-require "pundit"
 require "select2-rails"
 require "simple_form"
 require "godmin/admin_user"
@@ -14,6 +13,7 @@ require "godmin/helpers/modules"
 require "godmin/helpers/tables"
 require "godmin/helpers/translations"
 require "godmin/policy"
+require "godmin/policy_finder"
 require "godmin/rails"
 require "godmin/resolver"
 require "godmin/resource"
