+++ test/test_helper.rb
@@ -13,3 +13,9 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 if ActiveSupport::TestCase.method_defined?(:fixture_path=)
   ActiveSupport::TestCase.fixture_path = File.expand_path("../fixtures", __FILE__)
 end
+
+def mounted_as(namespace)
+  Godmin.mounted_as = namespace
+  yield
+  Godmin.mounted_as = nil
+end
